requirements.txt
langgraph>=0.2.40
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
openai>=1.50.0
chromadb>=0.5.5
tiktoken>=0.7.0
python-dotenv>=1.0.1
pydantic>=2.8.0
streamlit>=1.39.0
pytest>=8.3.0
pytest-mock>=3.14.0
rich>=13.8.0
